adding features, still does not converge while trainning on carla
