+++ vae_carla.py
@@ -6,19 +6,17 @@ from keras import layers
 import keras
 from keras.models import Model, load_model
 import numpy as np
-from config import IMG_SIZE
+from config import IMG_SIZE, mode, epochs, latent_dim, beta, scale, scale_r, lr, use_pretrained
+
 # filepath = "/home/gu/project/ppo/ppo_carla/models/carla/carla_vae_model_beta_3_r_1100.hdf5"
 batch_size = 100
 from keras.callbacks import ModelCheckpoint
 import vae_unit as vae_util
 
-use_pretrained = False
-epochs = 100
-latent_dim = 256
-beta = 3.0
-scale = 1.0
-scale_r = 1100
-filepath = "/home/gu/project/ppo/ppo_carla/models/carla/ld_{}_beta_{}_r_{}.hdf5".format(latent_dim, beta, scale_r)
+filepath = "/home/gu/project/ppo/ppo_carla/models/mnist/ld_{}_beta_{}_r_{}.hdf5".format(latent_dim, beta, scale_r)
+
+
+# filepath = "/home/gu/project/ppo/ppo_carla/models/mnist/ld_{}_beta_{}_r_{}.hdf5".format(latent_dim, beta, scale_r)
 
 def create_vae(latent_dim, return_kl_loss_op=False):
     '''
@@ -52,18 +50,15 @@ def create_vae(latent_dim, return_kl_loss_op=False):
         return model
 
 
-# Create plain VAE model and associated KL divergence loss operation
-vae, vae_kl_loss = create_vae(latent_dim, return_kl_loss_op=True)
-
-
 def vae_loss(x, t_decoded):
     '''Total loss for the plain VAE'''
-    return scale*K.mean(scale_r*reconstruction_loss(x, t_decoded) + beta * vae_kl_loss)  # adjust losss scale
-#
-#
+    return scale * K.mean(scale_r * reconstruction_loss(x, t_decoded) + beta * vae_kl_loss)  # adjust losss scale
+
+
 def reconstruction_loss(x, t_decoded):
     '''Reconstruction loss for the plain VAE'''
     return K.mean(K.square(x - t_decoded), axis=(1, 2, 3))  # average over images
+
 
 # def reconstruction_loss(x, t_decoded):
 #     '''Reconstruction loss for the plain VAE'''
@@ -71,36 +66,33 @@ def reconstruction_loss(x, t_decoded):
 #         K.batch_flatten(x),
 #         K.batch_flatten(t_decoded)), axis=-1)
 
-# x_train, x_test = load_carla_data(normalize=True, num=1000)
-x_train, x_test = load_carla_data(normalize=True)
-# (x_train,_), (x_test,_) = load_mnist_data(normalize=True)
-class SGDLearningRateTracker(keras.callbacks.Callback):
-    def on_epoch_end(self, epoch, logs={}):
-        optimizer = self.model.optimizer
-        # lr = K.eval(optimizer.lr * (1. / (1. + optimizer.decay * optimizer.iterations)))
-        # print('\nLR: {:.6f}\n'.format(lr))
-        print(K.eval(optimizer.iterations))
+
+if mode == "carla":
+    x_train, x_test = load_carla_data(normalize=True)
+else:
+    (x_train, _), (x_test, _) = load_mnist_data(normalize=True)
 
 # Keep only a single checkpoint, the best over test accuracy.
 checkpoint = ModelCheckpoint(filepath,
-                            monitor='val_loss',
-                            verbose=0,
-                            save_best_only=True,
-                            period=2,
-                            mode='max')
+                             monitor='val_loss',
+                             verbose=0,
+                             save_best_only=True,
+                             period=2,
+                             mode='max')
 
+# Create plain VAE model and associated KL divergence loss operation
+vae, vae_kl_loss = create_vae(latent_dim, return_kl_loss_op=True)
 if use_pretrained:
     vae.load_weights(filepath)
 else:
-    learning_rate = 0.001   # if we set lr to 0.005 network will blow up...that is...
-    # decay_rate = learning_rate / epochs
+    learning_rate = lr  # if we set lr to 0.005 network will blow up...that is...
     decay_rate = 1e-3
-    # decay_rate = 0
-
-    adam = keras.optimizers.Adam(lr=learning_rate, beta_1=0.9, beta_2=0.999, epsilon=None, decay=decay_rate, amsgrad=False)
+    adam = keras.optimizers.Adam(lr=learning_rate, beta_1=0.9, beta_2=0.999, epsilon=None, decay=decay_rate,
+                                 amsgrad=False)
     vae.compile(optimizer=adam, loss=vae_loss)
     history = vae.fit(x=x_train, y=x_train, epochs=epochs, batch_size=batch_size,
-            shuffle=True, validation_data=(x_test, x_test), verbose=2, callbacks=[checkpoint])# , SGDLearningRateTracker()])
+                      shuffle=True, validation_data=(x_test, x_test), verbose=2,
+                      callbacks=[checkpoint])  # , SGDLearningRateTracker()])
 
 
 def encode(model, images):
@@ -135,7 +127,3 @@ plot_image_rows([selected, selected_dec_vae],
                 ['Original images',
                  'Images generated by plain VAE'])
 plt.show()
-
-
-
-
