+++ vae_unit.py
@@ -225,7 +225,7 @@ def plot_image_rows(images_list, title_list):
     cols = len(images_list[0])
     # plt.figure(figsize=(cols, 3))
     def plot_image_row(images, title):
-        plt.figure(figsize=(cols, 3))
+        plt.figure(figsize=(cols, 2))
         plt.gcf().suptitle(title)
         for i, img in enumerate(images):
             plt.subplot(rows, cols, i + 1)
@@ -241,7 +241,6 @@ def plot_image_rows(images_list, title_list):
         plot_image_row(images, title)
 
 
-
 def plot_laplacian_variances(lvs_1, lvs_2, lvs_3, title):
     plt.hist(lvs_1, alpha=0.2, bins=50, label='Original images')
     plt.hist(lvs_2, alpha=0.2, bins=50, label='Images generated by plain VAE')
@@ -249,3 +248,11 @@ def plot_laplacian_variances(lvs_1, lvs_2, lvs_3, title):
     plt.xlabel('Laplacian variance')
     plt.title(title)
     plt.legend()
+
+
+class SGDLearningRateTracker(keras.callbacks.Callback):
+    def on_epoch_end(self, epoch, logs={}):
+        optimizer = self.model.optimizer
+        # lr = K.eval(optimizer.lr * (1. / (1. + optimizer.decay * optimizer.iterations)))
+        # print('\nLR: {:.6f}\n'.format(lr))
+        print(K.eval(optimizer.iterations))
